Mejora
Hice una separación de responsabilidades en el punto 1, además de que separé las funciones en sus debidos archivos.
