+++ Animales
@@ -1,25 +1,31 @@
 Leon africano
-Tigre de Bengala
-Oso panda gigante
-Gorila de montana
+Tigre siberiano
+Oso polar
+Elefante africano
+Jirafa reticulada
+Rinoceronte blanco
+Hipopotamo comun
+Gorila occidental de llanura
 Chimpance comun
 Orangutan de Borneo
-Elefante africano de sabana
-Rinoceronte blanco
-Jirafa reticulada
-Hipopotamo comun
 Lobo gris
-Colibri garganta de rubi
-Guacamayo azul y amarillo
-Pinguino emperador
-Foca leopardo
+Zorro rojo
+Lince iberico
+Puma
+Guepardo
+Leopardo de las nieves
+Oso pardo
 Ballena jorobada
-Delfin rosado del Amazonas
-Serpiente cascabel de diamante occidental
+Delfin nariz de botella
+Orca
+Morsa
+Foca comun
+Pinguino emperador
+Aguila real
+Condor andino
+Halcon peregrino
+Buho real
+Serpiente cascabel
 Cocodrilo del Nilo
-Salamandra tigre
-Rana flecha venenosa
-Arana viuda negra
-Escorpion emperador
 Mariposa monarca
-Pulga de gato
+Saltamontes migratorio
