+++ funciones.py
@@ -1,10 +1,10 @@
 import unicodedata
 import pickle
+import random
 
-def limpiarTexto(texto):
-    textoNorm = unicodedata.normalize('NFKD', texto)
-    textoAscii = textoNorm.encode('ascii', 'ignore').decode('ascii')
-    return textoAscii
+##################################################
+#Funciones de manipulación de datos
+##################################################
 
 def grabaTxt(archivoTxt,datos):
     """
@@ -48,7 +48,74 @@ def cargarPickle(nombreArchivo):
         print(f"Error al cargar con pickle: {e}")
         return []
 
-#Clase Animal
+##################################################
+# 2. Crear inventario
+##################################################
+
+def seleccionarAnimalesAleatorios(rutaArchivo, cantidad=20): #Lee cada linea del txt y toma 20 animales
+    with open(rutaArchivo, "r", encoding="utf-8") as f:
+        nombres = f.read().splitlines()
+    seleccionados = random.sample(nombres, cantidad)
+    print("Animales seleccionados para buscar en Wikipedia:") #Mostrar temporalmente los animales seleccionados en terminal
+    for nombre in seleccionados:
+        print(nombre)
+    return seleccionados
+
+def generarId(nombre, consecutivo): #Generar id único del nombre
+    nombre = nombre.lower()
+    primera = nombre[0]
+    ultima = nombre[-1]
+    return f"{primera}{ultima}{consecutivo:02d}" #Devuelve primer, última letra y consecutivo con dos dígitos
+
+def obtenerDatosAnimalGemini(model, nombreComun):
+    #Prompt utilizado para generar los datos del animal
+    prompt = (
+        f"Dame el nombre científico, el tipo de alimentación (solo responde 'carnívoro', 'herbívoro' u 'omnívoro') "
+        f"y una URL de imagen del animal '{nombreComun}'. No uses viñetas ni encabezados. Responde separado por saltos de línea.")
+    try:
+        response = model.generate_content(prompt)
+        datos = response.text.strip().split("\n")
+        nombreCientifico = datos[0].strip()
+        tipoAlimentacion = datos[1].strip().lower()
+        urlImagen = datos[2].strip()
+        return nombreCientifico, tipoAlimentacion, urlImagen
+    except Exception as e:
+        print(f"Error al obtener datos de {nombreComun}: {e}") #Puede dar error por límite de cuotas de Gemini
+        return "desconocido", "omnívoro", ""
+
+##################################################
+# 1. Obtener lista
+##################################################
+
+def peticionGeminiAnimales(numeroTotales, contenido):
+    mensaje = (
+        f"Genérame una lista de {numeroTotales} animales distintos, asegurándote de que cada uno sea específico y único." +
+        f"Usa exclusivamente los animales mencionados en el siguiente texto de Wikipedia: {contenido}" +
+        "Utiliza únicamente nombres comunes detallados (por ejemplo: 'Águila real', 'Zorro ártico', 'Delfín nariz de botella', 'Mariposa monarca')." +
+        "No uses nombres genéricos como 'Águila', 'Zorro' o 'Mariposa'." +
+        "Devuélvelos estrictamente en este formato: solo el nombre común, sin numeración ni negritas ni texto adicional." +
+        "Ejemplo de formato: León africano")
+    return mensaje
+
+def obtenerTextoAniLimpio(lineas):
+    listaAnimales = []
+    for linea in lineas:
+        lineaLimpia = linea.strip()
+        if lineaLimpia != "": #Verificar que la línea no esté vacía
+            listaAnimales.append(lineaLimpia) #Agregar la línea limpia a la lista
+    resultadoLimpio = "\n".join(listaAnimales)
+    print(resultadoLimpio)
+    return resultadoLimpio
+
+def limpiarTexto(texto): #Permite limpiar texto, permitiendo tildes y caracteres únicamente disponibles en el español.
+    textoNorm = unicodedata.normalize('NFKD', texto) #Convierte letras con tildes o símbolos especiales a una forma descompuesta.
+    textoAscii = textoNorm.encode('ascii', 'ignore').decode('ascii') #Elimina caracteres que no admite ASCII y .decode('ascii') devuelve texto plano
+    return textoAscii
+
+##################################################
+# Clase Animal
+##################################################
+
 class Animal:
     """
     Funcionamiento: La clase animal.
