+++ main.py
@@ -10,6 +10,7 @@ from funciones import *
 ##################################################
 # Ventanas Secundarias
 ##################################################
+
 def ventanaRetroalimentacion(mensaje):
     """
     Funcionamiento: 
@@ -68,37 +69,6 @@ def ventanaAprobacion(comandoAceptar):
 # 2. Crear inventario
 ##################################################
 
-def seleccionarAnimalesAleatorios(rutaArchivo, cantidad=20): #Lee cada linea del txt y toma 20 animales
-    with open(rutaArchivo, "r", encoding="utf-8") as f:
-        nombres = f.read().splitlines()
-    seleccionados = random.sample(nombres, cantidad)
-    print("Animales seleccionados para buscar en Wikipedia:") #Mostrar temporalmente los animales seleccionados en terminal
-    for nombre in seleccionados:
-        print(nombre)
-    return seleccionados
-
-def generarId(nombre, consecutivo): #Generar id único del nombre
-    nombre = nombre.lower()
-    primera = nombre[0]
-    ultima = nombre[-1]
-    return f"{primera}{ultima}{consecutivo:02d}" #Devuelve primer, última letra y consecutivo con dos dígitos
-
-def obtenerDatosAnimalGemini(model, nombreComun):
-    #Prompt utilizado para generar los datos del animal
-    prompt = (
-        f"Dame el nombre científico, el tipo de alimentación (solo responde 'carnívoro', 'herbívoro' u 'omnívoro') "
-        f"y una URL de imagen del animal '{nombreComun}'. No uses viñetas ni encabezados. Responde separado por saltos de línea.")
-    try:
-        response = model.generate_content(prompt)
-        datos = response.text.strip().split("\n")
-        nombreCientifico = datos[0].strip()
-        tipoAlimentacion = datos[1].strip().lower()
-        urlImagen = datos[2].strip()
-        return nombreCientifico, tipoAlimentacion, urlImagen
-    except Exception as e:
-        print(f"Error al obtener datos de {nombreComun}: {e}") #Puede dar error por límite de cuotas de Gemini
-        return "desconocido", "omnívoro", ""
-
 def crearObjetosAnimal(listaNombres):
     genai.configure(api_key="AIzaSyCj7ewGgJ0c7Cb_nHrHfzkN4lGDFlZ_iY0")
     model = genai.GenerativeModel(model_name="gemini-1.5-flash")
@@ -150,22 +120,12 @@ def obtenerLista(numeroTotales):
         # Configurar Gemini
         genai.configure(api_key="AIzaSyCj7ewGgJ0c7Cb_nHrHfzkN4lGDFlZ_iY0")
         model = genai.GenerativeModel(model_name = "gemini-1.5-flash")
-        response = model.generate_content(
-        f"Genérame una lista de {numeroTotales} animales distintos, asegurándote de que cada uno sea específico y único. " +
-        "Utiliza únicamente nombres comunes detallados (por ejemplo: 'Águila real', 'Zorro ártico', 'Delfín nariz de botella', 'Mariposa monarca'). " +
-        "No uses nombres genéricos como 'Águila', 'Zorro' o 'Mariposa'. " +
-        "Devuélvelos estrictamente en este formato: solo el nombre común, sin numeración ni negritas ni texto adicional. " +
-        "Ejemplo de formato: León africano") #Prompt modificado para generar nombres mas especificos
+        mensajePeticion = peticionGeminiAnimales(numeroTotales, contenido)
+        response = model.generate_content(mensajePeticion) #Prompt modificado para generar nombres mas especificos
         resultado = response.text
         resultado = limpiarTexto(resultado)
         lineas = resultado.splitlines()
-        listaAnimales = []
-        for linea in lineas:
-            lineaLimpia = linea.strip()
-            if lineaLimpia != "": #Verificar que la línea no esté vacía
-                listaAnimales.append(lineaLimpia) #Agregar la línea limpia a la lista
-        resultadoLimpio = "\n".join(listaAnimales)
-        print(resultadoLimpio)
+        resultadoLimpio = obtenerTextoAniLimpio(lineas)
         grabaTxt("Animales", resultadoLimpio)
         mensaje = f"Se ha creado el txt con los {numeroTotales} animales."
         ventanaConfirmacion(mensaje)
@@ -187,11 +147,13 @@ def ventanaObtenerLista():
     search = tk.Toplevel()
     search.title("Obtener animales")
     search.geometry("300x200")
+    etiqueta = tk.Label(search, text="Ingrese el número de animales:")
+    etiqueta.pack(pady=(10, 5)) #5 píxeles arriba y 10 abajo
     diccGlobal["numero"] = tk.StringVar()
     cantidad = tk.Entry(search, width=30, textvariable=diccGlobal["numero"])
-    cantidad.pack()
+    cantidad.pack(pady=5) #5 pixeles arriba y abajo
     searchButton = tk.Button(search, text="Buscar", width=30, command=lambda: validarObtenerLista(diccGlobal["numero"].get()))
-    searchButton.pack()
+    searchButton.pack(pady=(5, 10))
     search.mainloop()
 
 ##################################################
