--- prueba
@@ -1,30 +1,0 @@
-Leon
-Tigre
-Elefante
-Jirafa
-Ballena
-Delfin
-Serpiente
-Cocodrilo
-Tortuga
-Aguila
-Gorrion
-Murcielago
-Chimpance
-Gorila
-Orangutan
-Oso
-Zorro
-Conejo
-Caballo
-Vaca
-Cerdo
-Oveja
-Cabra
-Perro
-Gato
-Raton
-Lobo
-Hipopotamo
-Rinoceronte
-Pinguino
